+++ python/par_term_emu_core_rust/__init__.py
@@ -88,7 +88,7 @@ except ImportError:
     encode_client_message = None
     decode_client_message = None
 
-__version__ = "0.31.0"
+__version__ = "0.31.1"
 __all__ = [
     "AmbiguousWidth",
     "Attributes",